refactor utils logic to where its being used
