--- great_expectations/datasource/data_connector/__init__.py
@@ -1,4 +1,0 @@
-# isort:skip_file
-
-
-
