+++ great_expectations/datasource/fluent/data_connector/azure_blob_storage_data_connector.py
@@ -1,16 +1,13 @@
 from __future__ import annotations
 
 import logging
+import os
 import re
 from typing import TYPE_CHECKING, Callable, ClassVar, List, Optional, Type
 
-from great_expectations.compatibility import pydantic
+from great_expectations.compatibility import azure, pydantic
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.core.batch_spec import AzureBatchSpec, PathBatchSpec
-from great_expectations.datasource.data_connector.util import (
-    list_azure_keys,
-    sanitize_prefix,
-)
 from great_expectations.datasource.fluent.data_connector import (
     FilePathDataConnector,
 )
@@ -216,3 +213,61 @@ requires "file_path_template_map_fn: Callable" to be set.
         }
 
         return self._file_path_template_map_fn(**template_arguments)
+
+
+def sanitize_prefix(text: str) -> str:
+    """
+    Takes in a given user-prefix and cleans it to work with file-system traversal methods
+    (i.e. add '/' to the end of a string meant to represent a directory)
+    """
+    _, ext = os.path.splitext(text)  # noqa: PTH122
+    if ext:
+        # Provided prefix is a filename so no adjustment is necessary
+        return text
+
+    # Provided prefix is a directory (so we want to ensure we append it with '/')
+    return os.path.join(text, "")  # noqa: PTH118
+
+
+def list_azure_keys(
+    azure_client: azure.BlobServiceClient,
+    query_options: dict,
+    recursive: bool = False,
+) -> List[str]:
+    """
+    Utilizes the Azure Blob Storage connection object to retrieve blob names based on user-provided criteria.
+
+    For InferredAssetAzureDataConnector, we take container and name_starts_with and search for files using RegEx at and below the level
+    specified by those parameters. However, for ConfiguredAssetAzureDataConnector, we take container and name_starts_with and
+    search for files using RegEx only at the level specified by that bucket and prefix.
+
+    This restriction for the ConfiguredAssetAzureDataConnector is needed, because paths on Azure are comprised not only the leaf file name
+    but the full path that includes both the prefix and the file name.  Otherwise, in the situations where multiple data assets
+    share levels of a directory tree, matching files to data assets will not be possible, due to the path ambiguity.
+
+    Args:
+        azure_client (BlobServiceClient): Azure connnection object responsible for accessing container
+        query_options (dict): Azure query attributes ("container", "name_starts_with", "delimiter")
+        recursive (bool): True for InferredAssetAzureDataConnector and False for ConfiguredAssetAzureDataConnector (see above)
+
+    Returns:
+        List of keys representing Azure file paths (as filtered by the query_options dict)
+    """  # noqa: E501
+    container: str = query_options["container"]
+    container_client: azure.ContainerClient = azure_client.get_container_client(container=container)
+
+    path_list: List[str] = []
+
+    def _walk_blob_hierarchy(name_starts_with: str) -> None:
+        for item in container_client.walk_blobs(name_starts_with=name_starts_with):
+            if isinstance(item, azure.BlobPrefix):
+                if recursive:
+                    _walk_blob_hierarchy(name_starts_with=item.name)
+
+            else:
+                path_list.append(item.name)
+
+    name_starts_with: str = query_options["name_starts_with"]
+    _walk_blob_hierarchy(name_starts_with)
+
+    return path_list
