+++ tests/datasource/data_connector/test_data_connector_util.py
@@ -2,247 +2,16 @@ from unittest import mock
 
 import pytest
 
-import great_expectations.exceptions.exceptions as gx_exceptions
 from great_expectations.compatibility import google
-from great_expectations.core.batch import BatchRequest, IDDict, LegacyBatchDefinition
-
-# noinspection PyProtectedMember
-from great_expectations.datasource.data_connector.util import (
+from great_expectations.core.batch import IDDict, LegacyBatchDefinition
+from great_expectations.datasource.fluent.data_connector.file_path_data_connector import (
     _invert_regex_to_data_reference_template,
-    batch_definition_matches_batch_request,
-    build_sorters_from_config,
     convert_batch_identifiers_to_data_reference_string_using_regex,
-    convert_data_reference_string_to_batch_identifiers_using_regex,
-    list_gcs_keys,
     map_batch_definition_to_data_reference_string_using_regex,
-    map_data_reference_string_to_batch_definition_list_using_regex,
 )
-
-
-@pytest.mark.unit
-def test_batch_definition_matches_batch_request():
-    my_batch_definition = LegacyBatchDefinition(
-        datasource_name="test_environment",
-        data_connector_name="general_filesystem_data_connector",
-        data_asset_name="TestFiles",
-        batch_identifiers=IDDict({"name": "eugene", "timestamp": "20200809", "price": "1500"}),
-    )
-
-    # fully matching_batch_request
-    my_batch_request = BatchRequest(
-        datasource_name="test_environment",
-        data_connector_name="general_filesystem_data_connector",
-        data_asset_name="TestFiles",
-        data_connector_query=None,
-    )
-    assert batch_definition_matches_batch_request(my_batch_definition, my_batch_request) is True
-
-    # execution environment doesn't match
-    my_batch_request = BatchRequest(
-        datasource_name="i_dont_match",
-        data_connector_name="general_filesystem_data_connector",
-        data_asset_name="TestFiles",
-        data_connector_query=None,
-    )
-    assert batch_definition_matches_batch_request(my_batch_definition, my_batch_request) is False
-
-    # data_connector_name doesn't match
-    my_batch_request = BatchRequest(
-        datasource_name="test_environment",
-        data_connector_name="i_dont_match",
-        data_asset_name="TestFiles",
-        data_connector_query=None,
-    )
-    assert batch_definition_matches_batch_request(my_batch_definition, my_batch_request) is False
-
-    # data_asset_name doesn't match
-    my_batch_request = BatchRequest(
-        datasource_name="test_environment",
-        data_connector_name="general_filesystem_data_connector",
-        data_asset_name="i_dont_match",
-        data_connector_query=None,
-    )
-    assert batch_definition_matches_batch_request(my_batch_definition, my_batch_request) is False
-
-    # batch_request.data_connector_query.batch_filter_parameters is not dict
-    my_batch_request = BatchRequest(
-        datasource_name="test_environment",
-        data_connector_name="general_filesystem_data_connector",
-        data_asset_name="TestFiles",
-        data_connector_query={"batch_filter_parameters": 1},
-    )
-    assert batch_definition_matches_batch_request(my_batch_definition, my_batch_request) is False
-
-    # batch_identifiers do not match batch_definition.batch_identifiers
-    my_batch_request = BatchRequest(
-        datasource_name="test_environment",
-        data_connector_name="general_filesystem_data_connector",
-        data_asset_name="TestFiles",
-        data_connector_query={"batch_filter_parameters": {"i": "wont_work"}},
-    )
-    assert batch_definition_matches_batch_request(my_batch_definition, my_batch_request) is False
-
-
-@pytest.mark.unit
-def test_map_data_reference_string_to_batch_definition_list_using_regex():
-    # regex_pattern does not match --> None
-    data_reference = "alex_20200809_1000.csv"
-    regex_pattern = r"^(.+)_____________\.csv$"
-    group_names = ["name", "timestamp", "price"]
-    returned_batch_def_list = map_data_reference_string_to_batch_definition_list_using_regex(
-        datasource_name="test_datasource",
-        data_connector_name="test_data_connector",
-        data_asset_name=None,
-        data_reference=data_reference,
-        regex_pattern=regex_pattern,
-        group_names=group_names,
-    )
-    assert returned_batch_def_list is None
-
-    # no data_asset_name configured --> DEFAULT_ASSET_NAME
-    data_reference = "alex_20200809_1000.csv"
-    regex_pattern = r"^(.+)_(\d+)_(\d+)\.csv$"
-    group_names = ["name", "timestamp", "price"]
-    returned_batch_def_list = map_data_reference_string_to_batch_definition_list_using_regex(
-        datasource_name="test_datasource",
-        data_connector_name="test_data_connector",
-        data_asset_name=None,
-        data_reference=data_reference,
-        regex_pattern=regex_pattern,
-        group_names=group_names,
-    )
-    assert returned_batch_def_list == [
-        LegacyBatchDefinition(
-            datasource_name="test_datasource",
-            data_connector_name="test_data_connector",
-            data_asset_name="DEFAULT_ASSET_NAME",
-            batch_identifiers=IDDict(
-                {
-                    "name": "alex",
-                    "timestamp": "20200809",
-                    "price": "1000",
-                }
-            ),
-        )
-    ]
-
-    # data_asset_name configured
-    returned_batch_def_list = map_data_reference_string_to_batch_definition_list_using_regex(
-        datasource_name="test_datasource",
-        data_connector_name="test_data_connector",
-        data_asset_name="test_data_asset",
-        data_reference=data_reference,
-        regex_pattern=regex_pattern,
-        group_names=group_names,
-    )
-    assert returned_batch_def_list == [
-        LegacyBatchDefinition(
-            datasource_name="test_datasource",
-            data_connector_name="test_data_connector",
-            data_asset_name="test_data_asset",
-            batch_identifiers=IDDict(
-                {
-                    "name": "alex",
-                    "timestamp": "20200809",
-                    "price": "1000",
-                }
-            ),
-        )
-    ]
-
-
-@pytest.mark.unit
-def test_convert_data_reference_string_to_batch_identifiers_using_regex():
-    data_reference = "alex_20200809_1000.csv"
-    pattern = r"^(.+)_(\d+)_(\d+)\.csv$"
-    group_names = ["name", "timestamp", "price"]
-    assert convert_data_reference_string_to_batch_identifiers_using_regex(
-        data_reference=data_reference, regex_pattern=pattern, group_names=group_names
-    ) == (
-        "DEFAULT_ASSET_NAME",
-        IDDict(
-            {
-                "name": "alex",
-                "timestamp": "20200809",
-                "price": "1000",
-            }
-        ),
-    )
-
-    data_reference = "eugene_20200810_1500.csv"
-    pattern = r"^(.+)_(\d+)_(\d+)\.csv$"
-    group_names = ["name", "timestamp", "price"]
-    assert convert_data_reference_string_to_batch_identifiers_using_regex(
-        data_reference=data_reference, regex_pattern=pattern, group_names=group_names
-    ) == (
-        "DEFAULT_ASSET_NAME",
-        IDDict(
-            {
-                "name": "eugene",
-                "timestamp": "20200810",
-                "price": "1500",
-            }
-        ),
-    )
-    data_reference = "DOESNT_MATCH_CAPTURING_GROUPS.csv"
-    pattern = r"^(.+)_(\d+)_(\d+)\.csv$"
-    group_names = ["name", "timestamp", "price"]
-    assert (
-        convert_data_reference_string_to_batch_identifiers_using_regex(
-            data_reference=data_reference,
-            regex_pattern=pattern,
-            group_names=group_names,
-        )
-        is None
-    )
-
-    data_reference = "eugene_DOESNT_MATCH_ALL_CAPTURING_GROUPS_1500.csv"
-    pattern = r"^(.+)_(\d+)_(\d+)\.csv$"
-    group_names = ["name", "timestamp", "price"]
-    assert (
-        convert_data_reference_string_to_batch_identifiers_using_regex(
-            data_reference=data_reference,
-            regex_pattern=pattern,
-            group_names=group_names,
-        )
-        is None
-    )
-
-
-@pytest.mark.unit
-def test_convert_data_reference_string_to_batch_identifiers_using_regex_with_named_groups(
-    caplog,
-):
-    data_reference = "alex_20200809_1000.csv"
-    pattern = r"^(?P<name>.+)_(?P<timestamp>\d+)_(?P<price>\d+)\.csv$"
-
-    group_names = ["name", "timestamp", "price"]
-    assert convert_data_reference_string_to_batch_identifiers_using_regex(
-        data_reference=data_reference, regex_pattern=pattern, group_names=group_names
-    ) == (
-        "DEFAULT_ASSET_NAME",
-        IDDict(
-            {
-                "name": "alex",
-                "timestamp": "20200809",
-                "price": "1000",
-            }
-        ),
-    )
-
-    group_names = ["name", "timestamp", "cost"]  # Mismatch between "price" and "cost"!
-    assert convert_data_reference_string_to_batch_identifiers_using_regex(
-        data_reference=data_reference, regex_pattern=pattern, group_names=group_names
-    ) == (
-        "DEFAULT_ASSET_NAME",
-        IDDict(
-            {
-                "name": "alex",
-                "timestamp": "20200809",
-            }
-        ),
-    )
-    assert "The named group 'price' must explicitly be stated" in caplog.text
+from great_expectations.datasource.fluent.data_connector.google_cloud_storage_data_connector import (  # noqa: E501
+    list_gcs_keys,
+)
 
 
 @pytest.mark.unit
@@ -438,62 +207,6 @@ def test__invert_regex_to_data_reference_template():
         print(name)
         group_names = ["name", "timestamp"]
         _invert_regex_to_data_reference_template(regex_pattern=regex, group_names=group_names)
-
-
-@pytest.mark.unit
-def test_build_sorters_from_config_good_config():
-    sorters_config = [
-        {
-            "orderby": "desc",
-            "class_name": "NumericSorter",
-            "name": "price",
-        }
-    ]
-    sorters = build_sorters_from_config(sorters_config)
-    assert (
-        sorters.__repr__()
-        == """{'price': {
-  "name": "price",
-  "reverse": true,
-  "type": "NumericSorter"
-}}"""
-    )
-    assert (
-        sorters["price"].__repr__()
-        == """{
-  "name": "price",
-  "reverse": true,
-  "type": "NumericSorter"
-}"""
-    )
-    # no sorters by name of i_dont_exist
-    with pytest.raises(KeyError):
-        _ = sorters["i_dont_exist"]
-
-
-@pytest.mark.unit
-def test_build_sorters_from_config_bad_config():
-    # 1. class_name is bad
-    sorters_config = [
-        {
-            "orderby": "desc",
-            "class_name": "IDontExist",
-            "name": "price",
-        }
-    ]
-    with pytest.raises(gx_exceptions.PluginClassNotFoundError):
-        build_sorters_from_config(sorters_config)
-
-    # 2. orderby : not a real order
-    sorters_config = [
-        {
-            "orderby": "not_a_real_order",
-            "class_name": "NumericSorter",
-            "name": "price",
-        }
-    ]
-    with pytest.raises(gx_exceptions.SorterError):
-        build_sorters_from_config(sorters_config)
 
 
 @pytest.mark.skipif(
