+++ tests/datasource/fluent/data_asset/data_connector/test_s3_data_connector.py
@@ -11,13 +11,15 @@ from great_expectations.core import IDDict
 from great_expectations.core.batch import LegacyBatchDefinition
 from great_expectations.core.partitioners import FileNamePartitionerYearly
 from great_expectations.core.util import S3Url
-from great_expectations.datasource.data_connector.util import (
-    sanitize_prefix,
-    sanitize_prefix_for_gcs_and_s3,
-)
 from great_expectations.datasource.fluent import BatchRequest
 from great_expectations.datasource.fluent.data_connector import (
     S3DataConnector,
+)
+from great_expectations.datasource.fluent.data_connector.azure_blob_storage_data_connector import (
+    sanitize_prefix,
+)
+from great_expectations.datasource.fluent.data_connector.file_path_data_connector import (
+    sanitize_prefix_for_gcs_and_s3,
 )
 
 if TYPE_CHECKING:
